Grid refactor
* Grid has now been split into mutable and immutable versions that implement the respective interfaces
* The GridHelper was changed from a class to a trait
* Unit tests were updated for the new class structure

This refactor was done as it was becoming apparent that the original Grid was attempting to implement features that belong elsewhere in violation of the Single Responsibility Principle.  The fact that the accompanying unit test was so big was a code smell for that.  Refactoring mutability into another class and deferring concepts such as the solution not overlapping the original puzzle elsewhere allowed the tests to be far simpler whilst still providing full coverage.