Enums for row and column IDs
* Implemented RowIds and ColumnIds enums for enforcing valid row/column IDs
* Removed unnecessary input validations
* Removed no-longer necessary unit tests

Switching over to enums has made the code in most cases cleaner as a number of assert methods could be removed along with the tests that ensured that those assertions were met.  It also means that there are no circumstances under which an invalid row or column can be entered.

There are a few places where there's a little extra added complexity (having to extract the enum values for use as array keys for example) but the simplifications enabled elsewhere by the change outweigh this inconvenience.