OptionFinder refactor
Re-implemented the caching logic in the OptionFinder class so that columns and subgrids are only cached for the duration of a single run.  This ensures that it will continue to provide correct results with a mutable grid.

The Finder was also refactored to accept the grid when invoked, rather than when constructed